Fix for deleteing a user to skip trying to delete non-existant related models - for example for a never logged in user

+++ app/User.php
@@ -83,7 +83,9 @@ class User extends Model implements AuthenticatableContract, AuthorizableContrac
                         }
                     } else {
                         Log::info(" - removing user->$child");
-                        $user->$child->delete();
+                        if ($user->child) {
+                            $user->$child->delete();
+                        }
                     }
                 }
             }
